droidlogic-tv: refactor the logic of show hdmi source [1/1]
PD#SWPL-21676

Problem:
Enter hdmi source from aml MboxLauncher, or return to
MboxLauncher and then enter hdmi source again, tv can't
send key events to the playback, the active source is
internal address.

Solution:
The select TV internal address action is done in TvInputSesseion
onRelease method, but for the Launcher and LiveTv the sequence
can't be sure, and there could be a senario in which the new
session onSetMain and then the old one onRelease, so remove it.

Verify:
Marconi

Change-Id: Ie5fccb8bac5aa0c1a57522d6c7cdae090b3ac6f8
Signed-off-by: an.xi <an.xi@amlogic.com>