cec: solve the media play/pause key issue [1/1]
PD#SWPL-109168

Problem:
The source player could be paused and then resumed
instantly when pressing the pause/play key once.

Solution:
The release of the keyevent could be sent with
a different media keycode. The media keycode
should be same in an event of press and release.

Verify:
t7_an400

Change-Id: I6432c96f77872f2a8757aa9c71a49924b08e87f8
Signed-off-by: an.xi <an.xi@amlogic.com>